+++ VS/CPP/hli-stl/span.ipp
@@ -6,7 +6,7 @@ namespace hli
 {
 
 	template <class T>
-	using span = std::tuple<typename T * const, const size_t>;
+	using span = std::tuple<typename T * const, const int>;
 
 
 }
